requirements.txt
Django==1.8
Pillow==2.8.2
gunicorn==19.3.0
supervisor==3.1.3
mysqlclient==1.3.7
djangorestframework==3.3.1
django-cors-headers==1.1.0
django-filter==0.11.0
django-suit==0.2.15
requests==2.7
simplejson==3.7.3
redis==2.10.3
django-redis-cache==1.6.5
drf-extensions==0.2.8
newrelic==2.54.0.41
